Updated README. So long, and thanks for all the fish.
Sadly, the app hardly seems to work anymore. A dependency on Parse,
some CSS issues, and outdated frameworks that were bolted together
over a weekend does not a robust system make.

It was fun!

Leaving the repo as-is in case someone gets the spirit to bring the
thing up to date. Shutting down the Heroku dyno. The app should be
portable to any new dyno by creating new API keys for W3W and
Spotify.